add MultiCall to `CallHandler`
