+++ e2e/tests/contracts.rs
@@ -187,9 +187,7 @@ async fn test_multi_call_beginner() -> Result<()> {
     let call_handler_1 = contract_methods.get_single(7);
     let call_handler_2 = contract_methods.get_single(42);
 
-    let mut multi_call_handler = MultiContractCallHandler::new(wallet.clone());
-
-    multi_call_handler
+    let multi_call_handler = CallHandler::new_multi_call(wallet.clone())
         .add_call(call_handler_1)
         .add_call(call_handler_2);
 
@@ -227,9 +225,7 @@ async fn test_multi_call_pro() -> Result<()> {
     let call_handler_5 = contract_methods.get_array([7; 2]);
     let call_handler_6 = contract_methods.get_array([42; 2]);
 
-    let mut multi_call_handler = MultiContractCallHandler::new(wallet.clone());
-
-    multi_call_handler
+    let multi_call_handler = CallHandler::new_multi_call(wallet.clone())
         .add_call(call_handler_1)
         .add_call(call_handler_2)
         .add_call(call_handler_3)
@@ -347,9 +343,7 @@ async fn mult_call_has_same_estimated_and_used_gas() -> Result<()> {
     let call_handler_1 = contract_methods.initialize_counter(42);
     let call_handler_2 = contract_methods.get_array([42; 2]);
 
-    let mut multi_call_handler = MultiContractCallHandler::new(wallet.clone());
-
-    multi_call_handler
+    let multi_call_handler = CallHandler::new_multi_call(wallet.clone())
         .add_call(call_handler_1)
         .add_call(call_handler_2);
 
@@ -775,11 +769,11 @@ async fn test_output_variable_estimation_multicall() -> Result<()> {
     let amount = 1000;
     let total_amount = amount * NUM_OF_CALLS;
 
-    let mut multi_call_handler = MultiContractCallHandler::new(wallets[0].clone());
-    (0..NUM_OF_CALLS).for_each(|_| {
+    let mut multi_call_handler = CallHandler::new_multi_call(wallets[0].clone());
+    for _ in 0..NUM_OF_CALLS {
         let call_handler = contract_methods.mint_to_addresses(amount, addresses);
-        multi_call_handler.add_call(call_handler);
-    });
+        multi_call_handler = multi_call_handler.add_call(call_handler);
+    }
 
     wallets[0]
         .force_transfer_to_contract(
@@ -793,7 +787,7 @@ async fn test_output_variable_estimation_multicall() -> Result<()> {
 
     let base_layer_address = Bits256([1u8; 32]);
     let call_handler = contract_methods.send_message(base_layer_address, amount);
-    multi_call_handler.add_call(call_handler);
+    multi_call_handler = multi_call_handler.add_call(call_handler);
 
     let _ = multi_call_handler
         .estimate_tx_dependencies(None)
@@ -981,18 +975,18 @@ async fn test_output_variable_contract_id_estimation_multicall() -> Result<()> {
     let contract_methods = contract_caller_instance.methods();
 
     let mut multi_call_handler =
-        MultiContractCallHandler::new(wallet.clone()).with_tx_policies(Default::default());
+        CallHandler::new_multi_call(wallet.clone()).with_tx_policies(Default::default());
 
-    (0..3).for_each(|_| {
+    for _ in 0..3 {
         let call_handler = contract_methods.increment_from_contract(lib_contract_id, 42);
-        multi_call_handler.add_call(call_handler);
-    });
+        multi_call_handler = multi_call_handler.add_call(call_handler);
+    }
 
     // add call that does not need ContractId
     let contract_methods = contract_test_instance.methods();
     let call_handler = contract_methods.get(5, 6);
 
-    multi_call_handler.add_call(call_handler);
+    multi_call_handler = multi_call_handler.add_call(call_handler);
 
     let call_response = multi_call_handler
         .estimate_tx_dependencies(None)
@@ -1215,9 +1209,7 @@ async fn multi_call_from_calls_with_different_account_types() -> Result<()> {
     let call_handler_1 = contract_methods_wallet.initialize_counter(42);
     let call_handler_2 = contract_methods_predicate.get_array([42; 2]);
 
-    let mut multi_call_handler = MultiContractCallHandler::new(wallet);
-
-    multi_call_handler
+    let _multi_call_handler = CallHandler::new_multi_call(wallet)
         .add_call(call_handler_1)
         .add_call(call_handler_2);
 
@@ -1477,7 +1469,7 @@ async fn can_configure_decoding_of_contract_return() -> Result<()> {
     }
     {
         // Multi call: Will not work if max_tokens not big enough
-        MultiContractCallHandler::new(wallet.clone())
+        CallHandler::new_multi_call(wallet.clone())
         .add_call(methods.i_return_a_1k_el_array())
         .with_decoder_config(DecoderConfig { max_tokens: 100, ..Default::default() })
         .call::<([u8; 1000],)>().await.expect_err(
@@ -1486,7 +1478,7 @@ async fn can_configure_decoding_of_contract_return() -> Result<()> {
     }
     {
         // Multi call: Works when configured
-        MultiContractCallHandler::new(wallet.clone())
+        CallHandler::new_multi_call(wallet.clone())
             .add_call(methods.i_return_a_1k_el_array())
             .with_decoder_config(DecoderConfig {
                 max_tokens: 1001,
@@ -1526,9 +1518,7 @@ async fn test_contract_submit_and_response() -> Result<()> {
     let call_handler_1 = contract_methods.get_single(7);
     let call_handler_2 = contract_methods.get_single(42);
 
-    let mut multi_call_handler = MultiContractCallHandler::new(wallet.clone());
-
-    multi_call_handler
+    let multi_call_handler = CallHandler::new_multi_call(wallet.clone())
         .add_call(call_handler_1)
         .add_call(call_handler_2);
 
@@ -1573,9 +1563,7 @@ async fn test_heap_type_multicall() -> Result<()> {
         let call_handler_2 = contract_instance.methods().get_single(42);
         let call_handler_3 = contract_instance_2.methods().u8_in_vec(3);
 
-        let mut multi_call_handler = MultiContractCallHandler::new(wallet.clone());
-
-        multi_call_handler
+        let multi_call_handler = CallHandler::new_multi_call(wallet.clone())
             .add_call(call_handler_1)
             .add_call(call_handler_2)
             .add_call(call_handler_3);
@@ -1681,9 +1669,7 @@ async fn test_arguments_with_gas_forwarded() -> Result<()> {
         let call_handler_1 = contract_instance.methods().get_single(x);
         let call_handler_2 = contract_instance_2.methods().u32_vec(vec_input);
 
-        let mut multi_call_handler = MultiContractCallHandler::new(wallet.clone());
-
-        multi_call_handler
+        let multi_call_handler = CallHandler::new_multi_call(wallet.clone())
             .add_call(call_handler_1)
             .add_call(call_handler_2);
