+++ packages/fuels-programs/src/calls.rs
@@ -1,9 +1,10 @@
-mod callable;
 mod contract;
 pub mod receipt_parser;
 mod script;
+pub mod traits;
+mod tx_dependency_extension;
 pub mod utils;
 
-pub use callable::*;
 pub use contract::*;
 pub use script::*;
+pub use tx_dependency_extension::*;
