+++ packages/fuels-programs/src/calls/traits.rs
@@ -1,0 +1,7 @@
+mod buildable;
+mod extendable;
+mod parsable;
+
+pub use buildable::*;
+pub use extendable::*;
+pub use parsable::*;
